requirements.txt
#langchain==0.0.157
langchain_openai==0.3.1
langchain_core==0.3.31
streamlit==1.16.0
altair==4.2.2
python-dotenv==1.0.1

#langchain-community==0.3.15
langchain-huggingface==0.1.2
huggingface-hub==0.27.1
transformers==4.48.1
